requirements.txt
pypandoc
pyquery
dateparser
python-slugify
requests